Move traits file to lib
