+++ lib/traits.ts
